use detached context to fix #1059 #1059
It is important to use a detached context when creating shared transport connections that will outlive the initializing connection.

Otherwise, once the original initializing connection is cancelled, the transport connection will be cancelled as well.

This issue will be addressed in v5 by providing a transport session storage that also reduce the usage of globalDialerMap and other global variables.