Proper TCP Keep-Alive implementation.
This fix disables golang's Keep-Alive machanism (by setting -1 to dialer and listener) and controls it manually with setsockopt.

The default Keep-Alive initial interval is now 600 seconds (10 minutes), and interval between TCP keep-alive probes is not touched (75 seconds on Linux, OS default).

To completely disable Keep-Alive, set tcpKeepAliveInterval to -1.
Example:

    "streamSettings": {"sockopt": {"tcpKeepAliveInterval": -1}}

On Windows, Keep-Alive is only enabled, but the interval is not set.