Remove hysteria2
The hysteria2 fork used "github.com/JimmyHuang454/hysteria" removed the original license of hysteria2 (which is not allowed) and replaced it with GPL which conflicts with v2ray's license.

Given that the quality of the related code is very poor and it is not synchronized with the upstream, I deleted it.